refactor: move Redux store out of app routes and align React dependencies
- move Redux store/slices from route-scanned app/store to top-level store to stop expo-router route warnings
- update root layout and test imports to new store path
- fix habitSlice API import after folder move
- update react, react-dom, and react-test-renderer to 19.2.4 for peer compatibility
- update lockfile accordingly
