+++ Milestones/milestone_10/milestone10Project/package.json
@@ -36,8 +36,8 @@
     "expo-symbols": "~1.0.8",
     "expo-system-ui": "~6.0.9",
     "expo-web-browser": "~15.0.10",
-    "react": "19.1.0",
-    "react-dom": "19.1.0",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
     "react-native": "0.81.5",
     "react-native-gesture-handler": "~2.28.0",
     "react-native-reanimated": "~4.1.1",
@@ -58,7 +58,7 @@
     "jest": "^30.2.0",
     "jest-expo": "^55.0.9",
     "prettier": "^3.8.1",
-    "react-test-renderer": "^19.1.0",
+    "react-test-renderer": "^19.2.4",
     "typescript": "~5.9.2"
   },
   "private": true
